feat(hooks): :sparkles: add useAnimatedNumber
